Fix bug where server update has incorrect timeout
I'm specifying a timeout of 13ms, but occasionally  server.update(13) take 25-26 ms to return.

I'm certain the issue is with the hard-coded 25ms when Nio "freaks out".

I think the better solution would be to target a time of the minimum of  (25, timeout). I.e. if you're calling server.update(13), then don't sleep for 25 ms, aim for the method returning in 13!